example of embedded and script for toggling zoom
